Rearranged some classes fixed some errors
I created a player class and moved the existing player to
be ai player. I did this because there is a fair bit of
functionality that will be overlapping.

I added a TomaccoPile class to hold onto some more state
while the game is playing.

I started running files one by one in python and removed any
syntax errors that screamed out loud.