Fixed spacing. Fixed dealing.
I let some tabs get through, removed them.

The dealer was dealing the exact same cards to each player.
It's been changed to deal the cards properly. Now the cards
are removed from the deck as they are dealt out. This will
also be useful for drawing at the beginning of turns.